+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.0.3] - 2018-12-23
+### Changed
+- order of menu's is set by checking shortcuts
+- icons are now proper workspace icons
+
 ## [0.0.2] - 2018-12-22
 ### Changed
 - Merged quick screen switch and quick render menu into one
@@ -19,5 +24,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 <!--### Official Rigify Info-->
 
+[0.0.3]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.3
 [0.0.2]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.2
 [0.0.1]:https://github.com/schroef/QuickSwitch/releases/tag/v.0.0.1